requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
pymongo==4.6.1 
gunicorn==23.0.0
google-genai